admin access

+++ drill-and-practice/routes/controllers/topicController.js
@@ -1,34 +1,44 @@
 import * as topicService from "../../services/topicService.js";
 import * as questionService from "../../services/questionService.js";
+import * as optionService from "../../services/optionService.js";
 import {validasaur} from "../../deps.js";
 
 
-const createTopic = async ({request, response, render}) => {
-    const body = request.body({type: "form"});
-    const params = await body.value;
-
-    const tName = params.get("name");
-
-    const [passes, errors] = await validasaur.validate({value: tName}, {
-        value: [validasaur.required, validasaur.minLength(1)]
-    });
-
-    if (!passes) {
-        console.log('ERRORS:', errors)
-        render("topicsAll.eta", {topics: await topicService.showTopics(),
-            errors});
+const createTopic = async ({request, response, render, user}) => {
+    if (!user.admin) {
+        response.status = 401
     } else {
-        await topicService.createTopic(
-            1,
-            tName,
-        );
+        console.log('ONLY ADMIN SHOULD SEE THIS')
 
-        response.redirect("/topics");
+
+        const body = request.body({type: "form"});
+        const params = await body.value;
+
+        const tName = params.get("name");
+
+        const [passes, errors] = await validasaur.validate({value: tName}, {
+            value: [validasaur.required, validasaur.minLength(1)]
+        });
+
+        if (!passes) {
+            console.log('ERRORS:', errors)
+            render("topicsAll.eta", {topics: await topicService.showTopics(),
+                errors});
+        } else {
+            await topicService.createTopic(
+                user.id,
+                tName,
+            );
+
+            response.redirect("/topics");
+        }
     }
+
+
 };
 
-const showTopics = async ({request, response, render}) => {
-    render("topicsAll.eta", {topics: await topicService.showTopics()});
+const showTopics = async ({request, response, render, user}) => {
+    render("topicsAll.eta", {topics: await topicService.showTopics(), user});
 };
 
 const findTopicById = async ({request, response, params, render}) => {
@@ -39,10 +49,24 @@ const findTopicById = async ({request, response, params, render}) => {
         });
 };
 
-const deleteTopic = async ({request, response, params}) => {
-    await topicService.deleteTopic(params.id);
+const deleteTopic = async ({request, response, params, user}) => {
 
-    response.redirect("/topics");
+    if (!user.admin) {
+        response.status = 401
+    } else {
+        console.log('ONLY ADMIN SHOULD SEE THIS')
+
+        const topicQuestions = await questionService.findQuestions(params.id);
+
+        for (const q of topicQuestions) {
+            await optionService.deleteAllQuestionOptions(q.id)
+        }
+
+        await questionService.deleteAllTopicQuestions(params.id)
+        await topicService.deleteTopic(params.id);
+
+        response.redirect("/topics");
+    }
 };
